+++ Changelog.md
@@ -7,6 +7,7 @@
 - Remove immediate change of scene's cosms, since they shouldn't change until the Save Changes button is pressed.
 - DN in `@Check` is now localized if a skill or attribute.
 - Display message to player if the "Done" (end turn) button can't find your combatant.
+- Only allow the **UP** button to appear when the current Drama Card shows UP for the actor's faction.
 
 ### Bug Fixes